requirements.txt
attrs==24.2.0
aws-cdk-lib==2.166.0
aws-cdk.asset-awscli-v1==2.2.210
aws-cdk.asset-kubectl-v20==2.1.3
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==38.0.1
cattrs==24.1.2
constructs==10.4.2
exceptiongroup==1.2.2
importlib_resources==6.4.5
iniconfig==2.0.0
jsii==1.104.0
packaging==24.2
pluggy==1.5.0
publication==0.0.3
pytest==8.3.3
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.2
typeguard==2.13.3
typing_extensions==4.12.2
zipp==3.20.2
